Basic macOS service implementation
No visible feedback, unlike the share extension.
