+++ SendToKodi/SendToKodi.entitlements
@@ -8,5 +8,7 @@
 	<array>
 		<string>group.com.tangemann.sendtokodi</string>
 	</array>
+	<key>com.apple.security.network.client</key>
+	<true/>
 </dict>
 </plist>
