+++ SendToKodiAction/ShareViewController.swift
@@ -35,35 +35,14 @@ class ShareViewController: NSViewController {
         self.extensionContext!.cancelRequest(withError: NSError(domain: NSCocoaErrorDomain, code: NSUserCancelledError, userInfo: nil))
     }
     
-    func generateRequestDataFromUrl(_ url: URL) -> Data! {
-            return ("{\"jsonrpc\": \"2.0\", \"method\": \"Player.Open\", \"params\": " +
-                "{\"item\": {\"file\" : \"plugin://plugin.video.sendtokodi/?\(url)\" }}, \"id\" : \"0\"}")
-                .data(using: String.Encoding.utf8)
-    }
-    
     func sendRequestToKodi(_ url: URL) {
-        let session = URLSession.shared
-        let hostname = UserDefaults(suiteName: USER_DEFAULTS_SUITE)!.string(forKey: "kodi_hostname")!
-        
-        let requestData = self.generateRequestDataFromUrl(url)
-        
-        let request = NSMutableURLRequest(url: URL(string: "http://\(hostname):8080/jsonrpc")!)
-        request.httpMethod = "POST"
-        request.setValue("application/json",         forHTTPHeaderField: "Accept")
-        request.setValue("application/json",         forHTTPHeaderField: "Content-Type")
-        request.setValue(String(describing: requestData?.count), forHTTPHeaderField: "Content-Length")
-        request.httpBody = requestData
-        
-        let task = session.dataTask(with: request as URLRequest, completionHandler: {
-            data, response, error in
+        Utils.sendRequestToKodi(url, completionHandler: {
             
-            if error == nil {
-                self.extensionContext!.completeRequest(returningItems: nil, completionHandler: nil)
-            }
-            else {
-                self.extensionContext!.cancelRequest(withError: NSError(domain: NSCocoaErrorDomain, code: NSUserCancelledError, userInfo: nil))
-            }
-        }) 
-        task.resume()
+            self.extensionContext!.completeRequest(returningItems: nil, completionHandler: nil)
+            
+        }) { (Error) in
+            
+            self.extensionContext!.cancelRequest(withError: NSError(domain: NSCocoaErrorDomain, code: NSUserCancelledError, userInfo: nil))
+        }
     }
 }
